[fix] WebConfig Bean 에러 해결 #14
