+++ backend/src/main/java/com/coffee/backend/BackendApplication.java
@@ -8,11 +8,9 @@ import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfi
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@EnableAsync
 @RestController
 @SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 @EnableCaching
