+++ backend/src/main/java/com/coffee/backend/domain/auth/controller/AuthenticationPrincipalArgumentResolver.java
@@ -13,7 +13,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-@Component
+@Component("authenticationPrincipalArgumentResolver")
 @RequiredArgsConstructor
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
