+++ backend/src/main/java/com/coffee/backend/domain/storage/repository/UploadFileRepository.java
@@ -1,6 +1,6 @@
-package com.coffee.backend.domain.Storage.Repository;
+package com.coffee.backend.domain.storage.repository;
 
-import com.coffee.backend.domain.Storage.entity.UploadFile;
+import com.coffee.backend.domain.storage.entity.UploadFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UploadFileRepository extends JpaRepository<UploadFile, Long> {
