+++ backend/src/main/java/com/coffee/backend/domain/storage/service/StorageService.java
@@ -1,9 +1,9 @@
-package com.coffee.backend.domain.Storage.service;
+package com.coffee.backend.domain.storage.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.coffee.backend.domain.Storage.Repository.UploadFileRepository;
-import com.coffee.backend.domain.Storage.entity.UploadFile;
+import com.coffee.backend.domain.storage.repository.UploadFileRepository;
+import com.coffee.backend.domain.storage.entity.UploadFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
